feat(common): add blocksuite reader lib (#11955)
close CLOUD-202
