+++ tsconfig.json
@@ -126,6 +126,7 @@
     { "path": "./packages/common/graphql" },
     { "path": "./packages/common/infra" },
     { "path": "./packages/common/nbstore" },
+    { "path": "./packages/common/reader" },
     { "path": "./packages/common/y-octo/node" },
     { "path": "./packages/frontend/admin" },
     { "path": "./packages/frontend/apps/android" },
